pyproject.toml
[project]
name = "multimodal-fake-news"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "captum>=0.8.0",
    "datasets>=4.1.0",
    "firecrawl-py>=4.3.6",
    "google-generativeai>=0.8.5",
    "nltk>=3.9.1",
    "numpy>=1.26.4",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "scikit-learn>=1.3.0",
    "streamlit>=1.49.1",
    "torch>=2.8.0",
    "torch-geometric>=2.6.1",
    "torchaudio>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.56.1",
    "wandb>=0.21.4",
]
